+++ src/test/java/com/ejoongseok/wmslive/location/feature/AssignLocationLPNTest.java
@@ -1,21 +1,22 @@
 package com.ejoongseok.wmslive.location.feature;
 
-import com.ejoongseok.wmslive.inbound.domain.LPN;
 import com.ejoongseok.wmslive.inbound.domain.LPNRepository;
-import com.ejoongseok.wmslive.location.domain.Location;
 import com.ejoongseok.wmslive.location.domain.LocationRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.util.Assert;
 
 class AssignLocationLPNTest {
 
     private AssignLocationLPN assignLocationLPN;
+    private LocationRepository locationRepository;
+    private LPNRepository lpnRepository;
 
     @BeforeEach
     void setUp() {
-        assignLocationLPN = new AssignLocationLPN();
+        locationRepository = new LocationRepository();
+        lpnRepository = new LPNRepository();
+        assignLocationLPN = new AssignLocationLPN(locationRepository, lpnRepository);
     }
 
     @Test
@@ -36,22 +37,4 @@ class AssignLocationLPNTest {
 
     }
 
-    private class AssignLocationLPN {
-        private LocationRepository locationRepository;
-        private LPNRepository lpnRepository;
-
-        public void assignLocationLPN(final Request request) {
-            final Location location = locationRepository.getByLocationBarcode(request.locationBarcode);
-            final LPN lpn = lpnRepository.getByLpnBarcode(request.lpnBarcode);
-
-            location.assignLPN(lpn);
-        }
-
-        public record Request(String locationBarcode, String lpnBarcode) {
-            public Request {
-                Assert.hasText(locationBarcode, "로케이션 바코드는 필수입니다.");
-                Assert.hasText(lpnBarcode, "LPN 바코드는 필수입니다.");
-            }
-        }
-    }
 }
